+++ packages/main-ui/src/pages/surveys/models/attribute_group_info.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, Default)]
+pub struct AttributeGroupInfo {
+    pub name: String,
+    pub attribute: String,
+    pub rate: i64,
+}
