+++ packages/main-ui/src/pages/surveys/models/mod.rs
@@ -1,1 +1,3 @@
+pub mod attribute_combination;
+pub mod attribute_group_info;
 pub mod current_step;
